Extract AbstractArenaContext as the arena-agnostic base of TabArenaContext
Introduce `AbstractArenaContext` (ABC) holding the arena-agnostic core every arena
context shares: construction of the task/method metadata collections, the
`task_metadata` legacy view, `subset_predicates`, `compare` / `compare_per_dataset`,
`subset_results`, `make_experiment_batch_runner` / `_subset_dataset_fold_repeats`,
`methods` / `method_metadata` / `get_method_rename_map` /
`_method_rename_map_to_display_names`, and `leaderboard_to_website_format`.

Arena-specific behavior is now three hooks:
- `_resolve_task_metadata_preset(name)` — named preset -> TaskMetadataCollection
- `_resolve_methods_preset(name, *, include_unverified)` — named preset -> [MethodMetadata]
- `load_results_paper(...)` — baseline/reference results for `compare`; the base
  returns an empty frame (a self-contained arena has no external baselines), so a
  custom arena needs no override to get an offline leaderboard.

Class-level `SUBSET_PREDICATES` / `_default_subsets` keep their override semantics; the
base ships only the universal predicates (problem_type + lite).

`TabArenaContext(AbstractArenaContext)` keeps the TabArena v0.1 presets (now in the two
hooks), overrides `load_results_paper` with the paper-results downloader and
`SUBSET_PREDICATES` / `_default_subsets`, and retains the full paper workflow (HPO /
portfolio / plotting / repo generation). `BeyondArenaContext` is unchanged (still a
TabArenaContext subclass): MRO is BeyondArena -> TabArena -> AbstractArena -> ABC.

Adds tests for a minimal standalone AbstractArenaContext subclass (empty baselines,
universal predicates, make_experiment_batch_runner + subset, preset delegation).
